+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 * XRootD: update recipe from ALICE, don't overwrite version in defaults
 * pythia6: use version from SND@LHC
 * ofi: update version to v1.14.0
+* PHOTOSPP: change source repo and update version and recipe
 
 ### Removed